requirements.txt
streamlit
whisper
pandas
numpy
scikit-learn
sentence-transformers
groq
joblib
python-dotenv